post comments to github (#66)
note there are two comments on this PR b/c we run `npx autoblocks
testing exec` twice in our e2e tests (one for python and one for ts),
but this is pretty atypical. afaik all our users run one `npx autoblocks
testing exec` command per commit. the comments will be updated in-place
as new commits are pushed to the branch (so the comment only reflects
the latest run on the branch)

if there is no pull request associated with the commit, we comment on
the commit instead